# src/swagger/swagger-output.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Item Simulator REST APIs",
    "description": "Chapter 3, 개인과제 - 아이템 시뮬레이터 "
  },
  "servers": [
    {
      "url": "/api/",
      "description": ""
    }
  ],
...
  "paths": {
    "/sign-up": {
      "post": {
        "tags": [
          "Users: 회원관련"
        ],
        "summary": "회원가입 API",
        "description": "신규 유저를 추가한다.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signUp"
              }
            }
          }
        }
      }
    },
    "/sign-in": {
      "post": {
        "tags": [
          "Users: 회원관련"
        ],
        "summary": "로그인 API",
        "description": "유저 정보를 확인한다.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signIn"
              }
            }
          }
        }
      }
    },
    "/items": {
      "get": {
        "tags": [
          "Items: 아이템관련"
        ],
        "summary": "아이템 목록 조회 API",
        "description": "모든 아이템을 조회한다.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getItems"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/getItems"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items: 아이템관련"
        ],
        "summary": "아이템 생성 API",
        "description": "아이템을 생성한다.",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addItem"
              }
            }
          }
        }
      }
    },
    "/items/{itemCode}": {
      "get": {
        "tags": [
          "Items: 아이템관련"
        ],
        "summary": "아이템 상세 조회 API",
        "description": "특정 아이템을 조회한다.",
        "parameters": [
          {
            "name": "itemCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getItem"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/getItem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Items: 아이템관련"
        ],
        "summary": "아이템 수정 API",
        "description": "특정 아이템을 수정한다.",
        "parameters": [
          {
            "name": "itemCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/editItem"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "signUp": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@email.com"
          },
          "password": {
            "type": "string",
            "example": "qwer1234"
          },
          "name": {
            "type": "string",
            "example": "김유저"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ],
        "xml": {
          "name": "signUp"
        }
      },
      "signIn": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@email.com"
          },
          "password": {
            "type": "string",
            "example": "qwer1234"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "signIn"
        }
      },
      "getItems": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "itemCode": {
                  "type": "number",
                  "example": 8
                },
                "itemName": {
                  "type": "string",
                  "example": "망가진 노트북2"
                },
                "itemType": {
                  "type": "string",
                  "example": "ACCESSORY"
                },
                "itemStat": {
                  "type": "object",
                  "properties": {
                    "power": {
                      "type": "string",
                      "example": "10"
                    },
                    "health": {
                      "type": "string",
                      "example": "20"
                    }
                  },
                  "required": [
                    "power",
                    "health"
                  ]
                },
                "itemPrice": {
                  "type": "number",
                  "example": 500
                }
              },
              "required": [
                "itemCode",
                "itemName",
                "itemType",
                "itemStat",
                "itemPrice"
              ]
            }
          }
        },
        "required": [
          "data"
        ],
        "xml": {
          "name": "getItems"
        }
      },
      "getItem": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "itemCode": {
                "type": "number",
                "example": 8
              },
              "itemName": {
                "type": "string",
                "example": "망가진 노트북"
              },
              "itemType": {
                "type": "string",
                "example": "ACCESSORY"
              },
              "itemStat": {
                "type": "object",
                "properties": {
                  "power": {
                    "type": "string",
                    "example": "10"
                  },
                  "health": {
                    "type": "string",
                    "example": "20"
                  }
                },
                "required": [
                  "power",
                  "health"
                ]
              },
              "itemPrice": {
                "type": "number",
                "example": 500
              }
            },
            "required": [
              "itemCode",
              "itemName",
              "itemType",
              "itemStat",
              "itemPrice"
            ]
          }
        },
        "required": [
          "data"
        ],
        "xml": {
          "name": "getItem"
        }
      },
      "addItem": {
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string",
            "example": "대나무 회초리"
          },
          "itemType": {
            "type": "string",
            "example": "WEAPON"
          },
          "itemStat": {
            "type": "object",
            "properties": {
              "power": {
                "type": "string",
                "example": "10"
              },
              "health": {
                "type": "string",
                "example": "20"
              }
            },
            "required": [
              "power",
              "health"
            ]
          },
          "itemPrice": {
            "type": "number",
            "example": 500
          }
        },
        "required": [
          "itemName",
          "itemType",
          "itemStat",
          "itemPrice"
        ],
        "xml": {
          "name": "addItem"
        }
      },
      "editItem": {
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string",
            "example": "대나무 회초리"
          },
          "itemStat": {
            "type": "object",
            "properties": {
              "power": {
                "type": "string",
                "example": "10"
              },
              "health": {
                "type": "string",
                "example": "20"
              }
            },
            "required": [
              "power",
              "health"
            ]
          }
        },
        "required": [
          "itemName",
          "itemStat"
        ],
        "xml": {
          "name": "editItem"
        }
      },
...
    }
  }
}